Rainbow Set Information

A dataset containing information on the 18 images in the rainbow image set load_stim_rainbow(). These images are from Pixabay, all using the Pixabay License, so they are free for commercial and noncommercial use without attribution.

Usage

rainbow_info

Format

A data frame with 18 rows and 8 variables:

photo_name

the name of the phto inthe simulus set

photo_id

the numeric ID from Pixabay

photographer

the photographer's name

user_name

the photographer's pixabay username

user_id

the photographer's pixabay user ID

colour

the general colour of the image

type

whether the image is an animal, landscape, or flower

url

the pixabay URL for the image

Examples

head(rainbow_info)
#> # A tibble: 6 × 8
#>   photo_name          photo_id photographer user_name user_id colour type  url  
#>   <chr>                  <int> <chr>        <chr>       <int> <fct>  <fct> <chr>
#> 1 bled-2608425_640     2608425 David Mark   12019      1.20e4 purple land… http…
#> 2 bright-4642061_640   4642061 Nika Akin    nika_akin  1.35e7 purple anim… http…
#> 3 butterfly-142506_6…   142506 Garoch       garoch     1.03e4 blue   anim… http…
#> 4 cardinal-2872966_6…  2872966 Jill Wellin… jillwell…  3.34e5 red    anim… http…
#> 5 flower-2071522_640   2071522 NaturePhoto  natureph…  1.93e5 green  flow… http…
#> 6 flowers-3215188_640  3215188 CANDICE CAN… candiix    7.10e6 orange flow… http…